+++ .github/actions/ruby-publish-action/entrypoint.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+cd clients/ruby/
+make clean deps lint package
+mkdir -p ~/.gem/
+touch ~/.gem/credentials
+echo "---" >> ~/.gem/credentials
+echo ":rubygems_api_key: ${RUBYGEMS_TOKEN}" >> ~/.gem/credentials
+make publish