+++ .github/actions/ruby-release-patch-action/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+git config --global user.email "opensource@shinesolutions.com"
+git config --global user.name "Shine Open Source"
+cd clients/ruby/
+make release-patch